Create chat code
This code sets up an event listener on the input field, which listens for when the user presses the "Enter" key. When the user presses Enter, the code gets the input text, clears the input field, and appends the input text to the output field. The input and output fields are assumed to be elements in the HTML with the id "inputField" and "outputField", respectively.

You can use this code as a starting point for your chat application, and add additional functionality as needed.